[service] Unit tests and minor fixes
