+++ service/lib/agama/software/manager.rb
@@ -90,9 +90,6 @@ module Agama
         # patterns selected by user
         @user_patterns = []
         @selected_patterns_change_callbacks = []
-
-        update_issues
-
         on_progress_change { logger.info(progress.to_s) }
       end
